av_backend/requirements.dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --extra=dev --output-file=requirements.dev.txt pyproject.toml
#
anyascii==0.3.1
    # via wagtail
appnope==0.1.3
    # via
    #   ipykernel
    #   ipython
argon2-cffi==21.3.0
    # via notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.5.2
    # via django
asttokens==2.0.8
    # via stack-data
attrs==22.1.0
    # via jsonschema
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via
    #   nbconvert
    #   wagtail
bleach==5.0.1
    # via nbconvert
cachetools==5.2.0
    # via google-auth
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via argon2-cffi-bindings
charset-normalizer==2.1.1
    # via requests
debugpy==1.6.3
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
django==4.1.2
    # via
    #   av-backend (pyproject.toml)
    #   django-extensions
    #   django-filter
    #   django-model-utils
    #   django-modelcluster
    #   django-permissionedforms
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   wagtail
django-environ==0.9.0
    # via av-backend (pyproject.toml)
django-extensions==3.2.1
    # via av-backend (pyproject.toml)
django-filter==22.1
    # via wagtail
django-model-utils==4.2.0
    # via av-backend (pyproject.toml)
django-modelcluster==6.0
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-taggit==3.0.0
    # via wagtail
django-treebeard==4.5.1
    # via wagtail
djangorestframework==3.14.0
    # via wagtail
draftjs-exporter==2.1.7
    # via wagtail
entrypoints==0.4
    # via jupyter-client
et-xmlfile==1.1.0
    # via openpyxl
executing==1.1.1
    # via stack-data
fastjsonschema==2.16.2
    # via nbformat
google-api-core==2.10.2
    # via google-api-python-client
google-api-python-client==2.64.0
    # via av-backend (pyproject.toml)
google-auth==2.12.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.0
    # via
    #   av-backend (pyproject.toml)
    #   google-api-python-client
google-auth-oauthlib==0.5.3
    # via av-backend (pyproject.toml)
googleapis-common-protos==1.56.4
    # via google-api-core
html5lib==1.1
    # via wagtail
httplib2==0.20.4
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via requests
ipykernel==6.16.0
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   notebook
    #   qtconsole
ipython==8.5.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipython-genutils==0.2.0
    # via
    #   notebook
    #   qtconsole
ipywidgets==8.0.2
    # via jupyter
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via
    #   nbconvert
    #   notebook
jsonschema==4.16.0
    # via nbformat
jupyter==1.0.0
    # via av-backend (pyproject.toml)
jupyter-client==7.3.5
    # via
    #   ipykernel
    #   jupyter-console
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.4.4
    # via jupyter
jupyter-core==4.11.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-widgets==3.0.3
    # via ipywidgets
l18n==2021.3
    # via wagtail
markupsafe==2.1.1
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mistune==2.0.4
    # via nbconvert
nbclient==0.7.0
    # via nbconvert
nbconvert==7.2.1
    # via
    #   jupyter
    #   notebook
nbformat==5.6.1
    # via
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.6
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   notebook
notebook==6.4.12
    # via jupyter
oauthlib==3.2.1
    # via requests-oauthlib
openpyxl==3.0.10
    # via tablib
packaging==21.3
    # via
    #   ipykernel
    #   nbconvert
    #   qtpy
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.2.0
    # via wagtail
prometheus-client==0.14.1
    # via notebook
prompt-toolkit==3.0.31
    # via
    #   ipython
    #   jupyter-console
protobuf==4.21.7
    # via
    #   google-api-core
    #   googleapis-common-protos
psutil==5.9.2
    # via ipykernel
psycopg2-binary==2.9.3
    # via av-backend (pyproject.toml)
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pygments==2.13.0
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
    #   qtconsole
pyparsing==3.0.9
    # via
    #   httplib2
    #   packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via jupyter-client
pytz==2022.4
    # via
    #   django-modelcluster
    #   djangorestframework
    #   l18n
pyzmq==24.0.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   qtconsole
qtconsole==5.3.2
    # via jupyter
qtpy==2.2.1
    # via qtconsole
requests==2.28.1
    # via
    #   google-api-core
    #   requests-oauthlib
    #   wagtail
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
send2trash==1.8.0
    # via notebook
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   google-auth
    #   google-auth-httplib2
    #   html5lib
    #   l18n
    #   python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.3
    # via django
stack-data==0.5.1
    # via ipython
tablib[xls,xlsx]==3.2.1
    # via wagtail
telepath==0.3
    # via wagtail
terminado==0.16.0
    # via notebook
tinycss2==1.1.1
    # via nbconvert
tornado==6.2
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
traitlets==5.4.0
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.12
    # via requests
wagtail==4.0.2
    # via av-backend (pyproject.toml)
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   html5lib
    #   tinycss2
widgetsnbextension==4.0.3
    # via ipywidgets
willow==1.4.1
    # via wagtail
xlrd==2.0.1
    # via tablib
xlsxwriter==3.0.3
    # via wagtail
xlwt==1.3.0
    # via tablib
